Call stripe.createAccount always before sending tokens to Flex API
